==> scripts/7_PriorPredChecks_newModel.R <==
### 7_PriorPredChecks.R
# Ignore data and see what is generated by priors alone
# O.L. Pescott
# Dec. 2019/Jan. 2020
rm(list=ls())
######################################
library(R2jags)
library(dplyr)
######################################
## Helper functions
ilt <- function(x){ # inverse logit function
  exp(x)/(1+exp(x))
}
logit <- function(x){ # logit function
  log(x/(1 - x))
}
##

load(file = "data/Achi_mille_grassSamples_2020-01-06.Rdata") ## pre-processed
Achi_mill_PAN$date.x <- as.Date(as.character(Achi_mill_PAN$date.x), format = "%d/%m/%Y")
#load("W:/PYWELL_SHARED/Pywell Projects/BRC/_BRC_projects/NPMS/Analyses/2018 08 - Per species trend analyses/r_proj/NPMStrends/data/Achi_mille_grassSamples_20180920.Rdata")
head(Achi_mill_PAN); tail(Achi_mill_PAN); unique(Achi_mill_PAN$dominUnify)
# reorder grazing
unique(Achi_mill_PAN$grazing) # "High"     "Low"      "Moderate"
Achi_mill_PAN$grazing <- factor(Achi_mill_PAN$grazing, levels(Achi_mill_PAN$grazing)[c(2,3,1)])
levels(Achi_mill_PAN$grazing)
Achi_mill_PAN$grazing <- as.numeric(Achi_mill_PAN$grazing)
# simplify cutting/mowing text
levels(Achi_mill_PAN$cutting) <- "cut/mow"
# reorder survey levels
levels(Achi_mill_PAN$title)# "Indicator survey"  "Inventory survey"  "Wildflower survey"
Achi_mill_PAN$title <- factor(Achi_mill_PAN$title, levels(Achi_mill_PAN$title)[c(3,1,2)])
levels(Achi_mill_PAN$title)
Achi_mill_PAN$title <- as.numeric(Achi_mill_PAN$title)

# domin things
domins <- read.csv(file = "data/dominScores.csv", header = T, stringsAsFactors = F)
#domins <- read.csv("W:/PYWELL_SHARED/Pywell Projects/BRC/_BRC_projects/NPMS/Analyses/2018 08 - Per species trend analyses/r_proj/NPMStrends/data/dominScores.csv", header = T, stringsAsFactors = F)
# t.per <- c(0.001,0.01,0.03,0.05,0.1,0.25,0.33,0.50,0.75,0.95,0.99) these are the cutpoints used in Pescott et al. 2016

## Required data 1
#N = N # the total number of plots
#Y = Y # the total number of years
#n.Plot.pos = n.Plot.pos # the total number of positive cover observations across all plot visits (= samples in NPMS database terms)
#cpos.Cens = cpos.Cens # indicator (is censored?) -- can be T/F or 1/0, doesn't matter as long as is.logical(cpos.Cens) == T
## note that for the NPMS all non-zero observations are censored; this might not be true though if we combined with CS data or similar with more "precise" cover data
#cpos.Latent = cpos.Latent # just NA values for latent observations (unknown value underlying censored observation)
#lims = lims # the limits to the intervals used for censoring 
#(again, these should be of one type only, unless data collected under different schemes are combined) -- see Pescott et al. (2016)

## 1. Achillea millefolium in grassland samples
###############################################
# Change 25.03.2019 -- delete plots for which
# samples are entirely NA (i.e. no known states)
###############################################
library(dplyr)
#Achi_mill_PAN <- (Achi_mill_PAN %>% 
#                    group_by(plot_id) %>%
#                    filter(any(!is.na(dominUnify))) # resolves to true if there is at least one non-NA value within a group of plot samples
#)

# not sure it really matters, but sorting the data by year and then plot first may simplify downstream things
Achi_mill_PAN$year <- format(Achi_mill_PAN$date.x, "%Y") # create year column
Achi_mill_PAN <- Achi_mill_PAN[order(Achi_mill_PAN$year, Achi_mill_PAN$plot_id), ]
uniPlots <- unique(Achi_mill_PAN$plot_id) # unique plot IDs - 31/12/2018 = 933
# create unique plot index (useful for cross-referencing later on?)
plotIndex <- data.frame(plot = uniPlots, index = 1:length(uniPlots))
N <- length(uniPlots) # number of unique plots
# check that all years are in the range of 2015:System time
if ( max(unique(Achi_mill_PAN$year)) > format(Sys.time(), "%Y") ) {
  print ("error")
} else {
  print("OK")
}
# 
Y <- length(unique(format(as.Date(Achi_mill_PAN$date.x), "%Y"))) # number of years
# All Domin data, including zeros and NAs
yOrig <- Achi_mill_PAN$dominUnify#; hist(yOrig, breaks = 10)
# All grazing data
x2 <- Achi_mill_PAN$grazing
# All survey level data
x3 <- Achi_mill_PAN$title
# number of samples with presences
n.Plot.pos <- length(Achi_mill_PAN$dominUnify[Achi_mill_PAN$dominUnify !='0' & !is.na(Achi_mill_PAN$dominUnify)]) # 372
## checks on n.Plot.pos
head(Achi_mill_PAN[!is.na(Achi_mill_PAN$dominUnify) & Achi_mill_PAN$dominUnify !='0',]) # data frame
head(Achi_mill_PAN$dominUnify[Achi_mill_PAN$dominUnify !='0' & !is.na(Achi_mill_PAN$dominUnify)]) # vector
## looks OK

cpos.Cens <- rep(1, n.Plot.pos) # all samples censored
cpos.Latent <- rep(NA, n.Plot.pos) # all require estimation of latent state
t <- c(1e-4, 0.01, # cut points (note that these need to correspond to inits.fn and cPos.init below for JAGS model)
       0.01, 0.03,
       0.03, 0.05,
       0.05, 0.1,
       0.1, 0.25,
       0.25, 0.33,
       0.33, 0.5,
       0.5, 0.75,
       0.75, 0.95, 
       0.95, 0.9999)
t = matrix(t, nrow = 10, ncol = 2, byrow = TRUE)
tdf <- as.data.frame(t)
colnames(tdf) <- c('L','U') # 'L'ower and 'U'pper bounds of categories
tdf$int <- c(1,2,3,4,5,6,7,8,9,10); tdf

# positive rows only
spPos <- Achi_mill_PAN[Achi_mill_PAN$dominUnify !='0' & !is.na(Achi_mill_PAN$dominUnify),]; head(spPos)
# add row number to allow resorting after merge (merge with sort = F does not actually do what we want, could also used plyer::join)
spPos$indexPos <- 1:nrow(spPos)
spPos <- merge(spPos, tdf, by.x = "dominUnify", by.y = "int", all.x = T, all.y = F)
spPos <- spPos[order(spPos$indexPos),]
lims <- spPos[,c("L","U")] # has the lower/upper cutpoints for all intervals

## Required data 2
#plot = plot # an indicator linking a percentage cover observation to its parent (spatially unique) plot
#year = year # an indicator linking a percentage cover observation to its parent year
#V2 = V2 # the total number of visits (samples), irrespective of whether there is a postive cover for a species or not
#plotZ = plotZ # an indicator linking a visit to its parent (spatially unique) plot
#yearZ = yearZ # an indicator linking a visit to its parent year
#x = x # visit-level detection history (binary)

#plot <- match(spPos$plot_id, Achi_mill_PAN$plot_id) # an indicator linking a percentage cover observation to its parent (spatially unique) plot
spPos <- merge(spPos, plotIndex, by.x = "plot_id", by.y = "plot", all.x = T, all.y = F)
spPos <- spPos[order(spPos$indexPos),]
plot <- spPos$index
#yDat <- data.frame(index = 1:length(unique(format(Achi_mill_PAN$date.x, "%Y"))), year = unique(format(Achi_mill_PAN$date.x, "%Y"))) # not really needed
spPos$year <- as.factor(spPos$year)
levels(spPos$year) <- 1:length(unique(format(Achi_mill_PAN$date.x, "%Y")))
year <- spPos$year # an indicator linking a percentage cover observation to its parent year
## All visit information
V2 <- nrow(Achi_mill_PAN) # the total number of visits (samples), irrespective of whether there is a positive cover for a species or not
#plotZ <- match(Achi_mill_PAN$plot_id, uniPlots)  # an indicator linking a visit to its parent (spatially unique) plot
Achi_mill_PAN <- merge(Achi_mill_PAN, plotIndex, by.x = "plot_id", by.y = "plot", all.x = T, all.y = F)
Achi_mill_PAN <- Achi_mill_PAN[order(Achi_mill_PAN$year, Achi_mill_PAN$plot_id), ]
plotZ <- Achi_mill_PAN$index
Achi_mill_PAN$year <- as.factor(Achi_mill_PAN$year)
levels(Achi_mill_PAN$year) <- 1:length(unique(format(Achi_mill_PAN$date.x, "%Y")))
yearZ <- Achi_mill_PAN$year # an indicator linking a visit to its parent year
y <- y1 <-  numeric()
for (i in 1:nrow(Achi_mill_PAN)) {
  if(is.na(Achi_mill_PAN$dominUnify[i])){ 
    y<- NA
    #print(c(i,x)) # this was just for catching a bug
  } else if (Achi_mill_PAN$dominUnify[i]==0) { 
    y <- 0
    #print(c(i,x))
  } else {
    y <- 1
    #print(c(i,x))
  }
  y1 <- c(y1,y)
}
y <- y1 # visit-level detection history (binary)

####################
## Send data to JAGS
####################
# Data list for passing to JAGS
### For PPP ###
lims[,"L"] <- 0.0001
lims[,"U"] <- 0.9999
###############
Data <- list(N = N,
             Y = Y,
             n.Plot.pos = n.Plot.pos,
             cposCens = cpos.Cens, # indicator (is censored?)
             cposLatent = cpos.Latent, # NA values for latent observations
             lims = lims,
             cuts = c(1e-16,1e-4,0.01,0.03,0.05,0.1,0.25,0.33,0.5,0.75,0.95,0.9999),
             year = year,
             V2 = V2,
             plotZ = plotZ,
             yearZ = yearZ,
             #y = y,
             y =  rep(NA, length(y1)), # for PPP
             #x2 = x2, # grazing (with NAs)
             x2 = rep(NA, length(x2)), # for PPP
             #x2 = c(rep(x = 1:3, times = 788) , 1),
             #x3 = x3, # survey level
             x3 = rep(NA, length(x3)), # for PPP
             #yOrig = yOrig,
             yOrig = rep(NA, length(yOrig)), # for PPP
             y.ind = rep(0, length(yOrig)), # for PPP -- all values come from estimated cover distribution
             g2Levs = 3, 
             g3Levs = 3,
             pi = c(0.33,0.33,0.33)) 
###################################### END OF DATA PREP

###########################################
## Initialisation of values for JAGS chains
###########################################
# Initial parameter values for JAGS
zinit <- matrix(1, nrow = N, ncol = Y)
inits.fn <- function() list(z = zinit,
                            sdC = runif(1,0,10),
                            sd0 = runif(1,0,10),
                            sd.m = runif(1,0,10),
                            sd.g2 = runif(1,0,10),
                            sd.g3 = runif(1,0,10),
                            mean.m = runif(1,0,1),
                            mean.p = runif(1,0,1),
                            mean.g2 = runif(1,0,1),
                            mean.g3 = runif(1,0,1),
                            #mu = rep(0.5, Y),
                            phi = rep(2, Y),
                            y.hat = rep(runif(1,0,1), V2),
                            g1 = runif(1,0,3),
                            #cposLatent = c(0.001,0.025,0.04,0.075,0.175,0.29,0.375,0.625,0.85,0.975)[spPos$dominUnify]
                            ## for PPP
                            cposLatent = runif(n = n.Plot.pos, min = 0.0001, max = 0.999) # PPP
)
#for ref only
cPosInit <- c(0.001,0.025,0.04,0.075,0.175,0.29,0.375,0.625,0.85,0.975)[spPos$dominUnify]

######################################
## JAGS model
######################################
sink('scripts/JAGS/JAGS_mod_PPP_v2.txt')
cat("
model{
for (j in 1:Y){ # years
    # positive cover
    alpha[j] <- mu[j] * phi[j]
    beta[j] <- (1 - mu[j]) * phi[j]
    cPos[j] ~ dbeta(alpha[j], beta[j]) T(1e-4,0.9999) # random draw of cPos for that year, based on data-informed beta distribution
  for (i in 1:N){ # plots
      # ZI cover
      C[i,j] <- z[i,j] * cPos[j] # combine random draw of cPos for year j with estimated plot-specific occupancy
      
      # occupancy
      z[i,j] ~ dbern(psi[i,j]) # true PA state of a plot within a year depends on occupancy probability psi
      logit(psi[i,j]) <- m[j] # year random intercept on occupancy prob (same as Sparta)
      } # end of plots loop
} # end of years loop

## Derived values from state model
for (j in 1:Y){ # number of years
    annOcc[j] <- (sum(z[,j]))/N # annual occupancy
    mC[j] <- mean(C[,j]) # annual mean cover, including zeros
    mPsi[j] <- mean(psi[,j]) # annual mean occupancy prob
}
avgOcc <- mean(annOcc[]) # average annual occupancy

## Plot positive covers
for(k in 1:n.Plot.pos){ 
    cposCens[k] ~ dinterval(cposLatent[k], lims[k,])
    cposLatent[k] ~ dbeta(alpha[year[k]], beta[year[k]]) T(1e-4,0.9999) # recorded cover when present follows beta distribution
}

## Observation model for all plot visits ([within-year] detection within plots)
for (a in 1:V2){
    y[a] ~ dbern(py[a]) # detectability influences detection
    py[a] <- z[plotZ[a], yearZ[a]] * p.dec[a] # true state x detectability
    p.dec[a] <- min(max(1e-4, p.Dec[a]), 0.9999) # trick to stop numerical problems
    
    # detectability model: g1 is cover effect, g2 is grazing effect (with missing values), g3 is survey level (no missing values)
    logit(p.Dec[a]) <- g0 + g1*y.cov[a] + g3[x3[a]] + g2[x2[a]] 
    y.cov[a] <- yOrig[a]*y.ind[a] + y.int[a]*(1-y.ind[a]) # y.cov is a combination of original Domin scores, and predicted Domin scores where NA
    y.int[a] <- dinterval(y.hat[a], cuts)
    y.hat[a] ~ dbeta(alpha[yearZ[a]],beta[yearZ[a]])T(1e-4,0.9999)
    x2[a] ~ dcat(pi) # missing values for grazing info randomly assigned
    x3[a] ~ dcat(pi) # PPP only
    yOrig[a] ~ dunif(0,10) # PPP only
}

### Priors ###
## Intercept for state model for occupancy
mean.m ~ dbeta(1,1)T(1e-4,0.9999)
m[1] ~ dnorm(logit(mean.m), tau0)
tau0 <- 1/pow(sd0, 2)
sd0 ~ dt(0, 1, 3)T(0,)
for(j in 2:Y){
  m[j] ~ dnorm(m[j-1], tau.m)T(logit(1e-4), logit(0.9999)) # fixing tau.m (e.g. at 4) results in much smoother trend
}  
tau.m <- 1/pow(sd.m, 2)
sd.m ~ dt(0, 1, 3)T(0,)

## Cover (without random walk prior)
#for (j in 1:Y){
#phi[j] ~ dpar(1.5, 0.1) # alpha, c
#mu[j] ~ dbeta(1, 1)T(1e-4,0.9999)
#}
#### for more precise random walk on cover scale, we would need to know how to transform phi onto the logit scale ####
#
for (j in 1:Y){
phi[j] ~ dpar(1.5, 0.1) # alpha, c
}
mu[1] ~ dbeta(1, 1)T(1e-4,0.9999)
for (j in 2:Y){
  mu[j] <- ilogit(mInt[j]) # mean follows low-variation random walk on logit-normal scale (sd = 0.5, so tau = 4)
  mInt[j] ~ dnorm(logit(mu[j-1]), tauC)T(logit(1e-4), logit(0.9999)) # fixing tauC (e.g. at 4) results in much smoother trend
  #mInt[j] ~ dnorm(logit(mu[j-1]), 4)T(logit(1e-4), logit(0.9999)) # fixing tauC (e.g. at 4) results in much smoother trend
}
tauC <- 1/pow(sdC, 2)
sdC ~ dt(0, 1, 3)T(0,) #-- weaker, density mostly -10 - 10
#sdC ~ dt(0, 0.001, 1)T(0,) # bit too strong (see 7_PriorPredChecks.R script) -- normally most density -5 to 5

## Detection model
mean.p ~ dbeta(1,1)T(1e-4,0.9999) # broad intercept on prob scale
g0 <- logit(mean.p) # transformed # note that this requires tweak to initial values
g1 ~ dnorm(0, 0.1)T(0,) # moderate belief that this parameter should be positive, increase in cover increases detectability, all else being equal

# Hierarchical prior for levels of grazing
for (j in 1:g2Levs) { g2[j] ~ dnorm(g2mu, g2tau) } # shouldn't there be a zeroth level here as well?
mean.g2 ~ dbeta(1,1)T(1e-3,0.999) # between -7 and 7
g2mu <- logit(mean.g2)
g2tau <- 1/pow(sd.g2, 2)
sd.g2 ~ dt(0, 1, 3)T(0,)

# Hierarchical prior for survey levels
for (j in 1:g3Levs) { g3[j] ~ dnorm(g3mu, g3tau) }
mean.g3 ~ dbeta(1,1)T(1e-3,0.999) # between 
g3mu <- logit(mean.g3)
g3tau <- 1/pow(sd.g3, 2)
sd.g3 ~ dt(0, 1, 3)T(0,) # dt(mu,tau,k)

} ## END MODEL
", fill = TRUE)
sink()

jagsModel <- jags.model(file= 'scripts/JAGS/JAGS_mod_PPP_v2.txt', data = Data, inits = inits.fn, n.chains = 6, n.adapt= 100)

# Specify parameters for which posterior samples are saved
para.names <- c('mC', 'mPsi', 'mu', 'annOcc', 'avgOcc', 'cPos', 'g0', 'g1', 'g2', 'g3', 'alpha', 'beta')
#para.names <- c('mC')
# Continue the MCMC runs with sampling
#jagsModel <- update(jagsModel, n.iter = 500)
samples <- coda.samples(jagsModel, variable.names = para.names, thin = 5, n.iter = 1000)
## Inspect results
out <- summary(samples)
mu_se <- out$stat[,c(1,3)] #make columns for mean and se
#save(samples, file = "outputs/")
gelman.diag(samples)
plot(samples)

### Visualise ###
#library(tidybayes)
library(MCMCvis)
library(mcmcplots)
MCMCtrace(samples, 
          params = c('mu'),
          ISB = FALSE,
          iter = 500,
          ind = TRUE,
          pdf = FALSE)
mcmcplot(samples)
caterplot(mcmcout = samples, parms = 'cPos')
denplot(mcmcout = samples, parms = 'mu')
traplot(samples, parms = 'mu')
parcorplot(samples, parms = c('cPos', 'mu'))

#############
#### dt & rbeta stuff -- apparently just using abs doesn't cut it for truncated t distribution!
#############
# below is for dt(0, 0.01, 1)T(0,)
#x <- rnorm(n=1000, 0, sd = sqrt(1/0.01)) # tau = 0.1
#s <- rgamma(n=1000, shape = 1/2, rate = 1/2) # k = 1; shape = rate = k/2 = 1/2
#y <- x/sqrt(s)
#hist(abs(y), breaks = 1000, xlim = c(0,10))

# cf. to used in Kruschke (much lower density around 0-50 and much longer RHS tail): pg.496
#x <- abs(rnorm(n=1000, 0, sd = sqrt(1/0.001))) # tau = 0.001
#s <- rgamma(n=1000, shape = 2/2, rate = 2/2) # k = 2; shape = rate = k/2 = 1/2
#y <- x/sqrt(s)
#hist(y, breaks = 10000, xlim = c(0,500))

# Visualise priors for cPos
#pMu <- rbeta(n = 1000, shape1 = 1, shape2 = 1)
#pMu2 <- runif(n = 1000, min = 0, max = 1)
#pX <- abs(rnorm(n=1000, 0, sd = sqrt(1/0.1))) # 
#pS <- rgamma(n=1000, shape = 1/2, rate = 1/2) # k = 1; shape = rate = k/2 = 1/2
#pPhi <- pX/sqrt(pS)
#hist(pPhi)
#pA <- pMu * pPhi
#pB <- (1 - pMu) * pPhi
#hist(rbeta(n=1000, shape1 = pA, shape2 = pB))

library(crch) # for truncated students t
#hist(rtt(n = 1000, location = 0, scale = 0.1, df = 1, left = 0, right = Inf), breaks = 10000, xlim = c(0,500))
#hist(rtt(n = 1000, location = 0, scale = 0.001, df = 2, left = 0, right = Inf), breaks = 10000, xlim = c(0,500))
#hist(rtt(n = 1000, location = 0, scale = 0.01, df =1, left = 0, right = Inf), breaks = 10000, xlim = c(0,10))

# Visualise priors for cPos using crch
#pMu <- rbeta(n = 1000, shape1 = 1, shape2 = 1)
#pPhi <- rtt(n = 1000, location = 0, scale = 0.001, df = 2, left = 0, right = Inf) # Kruschke
##hist(pPhi)
#pA <- pMu * pPhi
#pB <- (1 - pMu) * pPhi
#hist(rbeta(n=1000, shape1 = pA, shape2 = pB))
#hist(logit(rbeta(n=1000, shape1 = pA, shape2 = pB)))

#pPhi <- rtt(n = 1000, location = 0, scale = 0.1, df = 1, left = 0, right = Inf)
#hist(pPhi)
#pA <- pMu * pPhi
#pB <- (1 - pMu) * pPhi
#hist(rbeta(n=1000, shape1 = pA, shape2 = pB))
#hist(logit(rbeta(n=1000, shape1 = pA, shape2 = pB)))

par(mfrow=c(2,2))
pMu <- rbeta(n = 1000, shape1 = 1, shape2 = 1)
pPhi <- rtt(n = 1000, location = 0, scale = 0.01, df = 1, left = 0, right = Inf) # Irvine etc.
#hist(pPhi, breaks = 1000)
pA <- pMu * pPhi
pB <- (1 - pMu) * pPhi
hist(rbeta(n=1000, shape1 = pA, shape2 = pB), main = "Draws from beta using half-Cauchy \nprior on tau")
hist(logit(rbeta(n=1000, shape1 = pA, shape2 = pB)), main = "Logit scale")

## Prior on phi much better as just normal
#pMu <- runif(n = 1000, min = 0, max = 1)
#pMu <- ilt(rnorm(n = 1000, 0,1))
#pPhi <- abs(rnorm(n = 1000, mean = 0, sd = 1)) #abs
#hist(pPhi, breaks = 1000)
#pA <- pMu * pPhi
#pB <- (1 - pMu) * pPhi
#hist(logit(rbeta(n=1000, shape1 = pA, shape2 = pB)))
#hist(rbeta(n=1000, shape1 = pA, shape2 = pB))

## Suggestions here https://mc-stan.org/docs/2_18/stan-users-guide/reparameterizations.html
###########################################################################################
# Only the below formulation result in uniform draws from rbeta
###########################################################################################
library(VGAM)
pPhi <- rpareto(1000, scale = 1.5, shape = 0.1) # scale = alpha, scale = k
pMu <- rbeta(1000,1,1)
pA <- pMu * pPhi
pB <- (1 - pMu) * pPhi
hist(rbeta(n=1000, shape1 = pA, shape2 = pB), main = "Draws from beta using Pareto \nprior on tau")
hist(logit(rbeta(n=1000, shape1 = pA, shape2 = pB)), main = "Logit scale")


### Priors ###
library(crch)
library(VGAM)
## Helper functions again
ilt <- function(x){ # inverse logit function
  exp(x)/(1+exp(x))
}
logit <- function(x){ # logit function
  log(x/(1 - x))
}
##
## Intercept for state model for occupancy
### JAGS CODE
#mean.m[1] ~ dbeta(1,1)T(1e-4,0.9999)
#for(j in 2:Y){
#  m[j] ~ dnorm(logit(mean.m[j-1]), tau.m)
#}  
#tau.m <- 1/pow(sd.mA, 2)
#sd.mA <- sd.m + 0.1 # see Kruschke page 486 (don't want shrinkage to be too strong when data sparse)
#sd.m ~ dt(0, 0.1, 1)T(0,)
###
par(mfrow=c(1,2))
sD <- rtt(n = 1000, location = 0, scale = 0.01, df = 1, left = 0, right = Inf) # Irvine etc.
mU <- logit(rbeta(1000,1,1))
hist(rnorm(n = 1000, mean = mU, sd = sD))
hist(ilt(rnorm(n = 1000, mean = mU, sd = sD)))

# SD + 0.1: see Kruschke page 486 (don't want shrinkage to be too strong when data sparse)
par(mfrow=c(1,2))
sD <- rtt(n = 1000, location = 0, scale = 0.01, df = 1, left = 0, right = Inf) # Irvine etc.
mU <- logit(rbeta(1000,1,1))
hist(rnorm(n = 1000, mean = mU, sd = sD+0.1))
hist(ilt(rnorm(n = 1000, mean = mU, sd = sD+0.1)))

#####################
### JAGS CODE
#mean.m ~ dbeta(1,1)T(1e-4,0.9999)
#m[1] ~ dnorm(logit(mean.m), tau0)
#tau0 <- 1/pow(sd0, 2)
#sd0 ~ dt(0, 0.01, 1)T(0,)
###
par(mfrow=c(3,3))
# half Cauchy, scale = 0.1
sD <- rtt(n = 1000, location = 0, scale = 0.1, df = 1, left = 0, right = Inf) # Irvine etc.
mU <- logit(rbeta(1000,1,1))
hist(rnorm(n = 1000, mean = mU, sd = sD))
hist(log(rnorm(n = 1000, mean = mU, sd = sD)))
hist(ilt(rnorm(n = 1000, mean = mU, sd = sD)))
# half Cauchy, scale = 0.001
sD <- rtt(n = 1000, location = 0, scale = 0.001, df = 1, left = 0, right = Inf) # Irvine etc.
mU <- logit(rbeta(1000,1,1))
hist(rnorm(n = 1000, mean = mU, sd = sD), breaks = 500)
hist(log(rnorm(n = 1000, mean = mU, sd = sD)))
hist(ilt(rnorm(n = 1000, mean = mU, sd = sD)))
# pareto
library(VGAM)
sDp <- rpareto(1000, scale = 1.5, shape = 0.1) # scale = alpha, scale = k
mUp <- logit(rbeta(1000,1,1))
hist(rnorm(n = 1000, mean = mUp, sd = sDp))
hist(log(rnorm(n = 1000, mean = mUp, sd = sDp)))
hist(ilt(rnorm(n = 1000, mean = mUp, sd = sDp)))

##
# half Cauchy, scale = 1, df = 3
# thinner tails, recommended by Aki Vehtari when data are sparse
sD <- rtt(n = 1000, location = 0, scale = 1, df = 3, left = 0, right = Inf) # Irvine etc.
mU <- logit(rbeta(1000,1,1))
hist(rnorm(n = 1000, mean = mU, sd = sD), breaks = 500)
#hist(log(rnorm(n = 1000, mean = mU, sd = sD)))
hist(ilt(rnorm(n = 1000, mean = mU, sd = sD)))


## Helper functions
ilt <- function(x){ # inverse logit function
  exp(x)/(1+exp(x))
}
logit <- function(x){ # logit function
  log(x/(1 - x))
}
##
